Update Evolution Database FichaManutencaoRelatorio Tables

+++ Hydra.Such.Portal/Hydra.Such.Data.Evolution/DatabaseReference/FichaManutencaoRelatorioTestesQualitativos.cs
@@ -8,7 +8,7 @@ namespace Hydra.Such.Data.Evolution.DatabaseReference
         public int Id { get; set; }
         public int IdRelatorio { get; set; }
         public int IdTesteQualitativos { get; set; }
-        public byte ResultadoRotina { get; set; }
+        public int ResultadoRotina { get; set; }
         public string Observacoes { get; set; }
     }
 }
